Move `_retrieve_scores` to `.endpoints`, also create `.const`
An oversight when I was working on v2.0.0

The method was only in the main class in the first place because this whole module used to be condensed into a class. `_retrieve_scores` isn't actually needed in the `MALAffinity` class, it's just needed to retrieve the scores off MAL. With v2.0.0, this can easily be moved out into its own file.

`const.py` is also created in case consts need to be stored, i.e. endpoint urls for services.

This also makes it easier and more convenient for #1 to happen, if possible (with a bit of tweaking).